Refactor move virtualenv into modules.pip
