+++ tests/operations/pip.virtualenv/add_with_options.json
