+++ tests/operations/pip.virtualenv/already_absent.json
