+++ tests/operations/pip.virtualenv/already_present.json
