+++ tests/operations/pip.virtualenv/delete_virtualenv.json
